+++ src/client/ui/app.tsx
@@ -1,6 +1,6 @@
 import React from "@rbxts/react";
 
-import Layer from "./components/primitive/layer";
+import { Layer } from "./components/primitive";
 
 export function App(): React.ReactNode {
 	return (
