+++ src/client/ui/components/primitive/canvas-group.tsx
@@ -19,22 +19,22 @@ export type CanvasGroupProps = FrameProps<CanvasGroup>;
  *
  * @see https://developer.roblox.com/en-us/api-reference/class/CanvasGroup
  */
-const CanvasGroup = forwardRef((props: Readonly<CanvasGroupProps>, ref: React.Ref<CanvasGroup>) => {
-	const { CornerRadius, Native, children } = props;
+export const CanvasGroup = forwardRef(
+	(props: Readonly<CanvasGroupProps>, ref: React.Ref<CanvasGroup>) => {
+		const { CornerRadius, Native, children } = props;
 
-	return (
-		<canvasgroup
-			ref={ref}
-			AnchorPoint={new Vector2(0.5, 0.5)}
-			BackgroundTransparency={1}
-			BorderSizePixel={0}
-			Position={new UDim2(0.5, 0, 0.5, 0)}
-			{...Native}
-		>
-			{CornerRadius ? <uicorner key="corner" CornerRadius={CornerRadius} /> : undefined}
-			{children}
-		</canvasgroup>
-	);
-});
-
-export default CanvasGroup;
+		return (
+			<canvasgroup
+				ref={ref}
+				AnchorPoint={new Vector2(0.5, 0.5)}
+				BackgroundTransparency={1}
+				BorderSizePixel={0}
+				Position={new UDim2(0.5, 0, 0.5, 0)}
+				{...Native}
+			>
+				{CornerRadius ? <uicorner key="corner" CornerRadius={CornerRadius} /> : undefined}
+				{children}
+			</canvasgroup>
+		);
+	},
+);
