+++ src/client/ui/components/primitive/group.tsx
@@ -22,19 +22,19 @@ interface GroupProps extends React.PropsWithChildren {
  *
  * @component
  */
-const Group = forwardRef(({ Native, children }: Readonly<GroupProps>, ref: React.Ref<Frame>) => {
-	return (
-		<frame
-			ref={ref}
-			AnchorPoint={new Vector2(0.5, 0.5)}
-			BackgroundTransparency={1}
-			Position={new UDim2(0.5, 0, 0.5, 0)}
-			Size={new UDim2(1, 0, 1, 0)}
-			{...Native}
-		>
-			{children}
-		</frame>
-	);
-});
-
-export default Group;
+export const Group = forwardRef(
+	({ Native, children }: Readonly<GroupProps>, ref: React.Ref<Frame>) => {
+		return (
+			<frame
+				ref={ref}
+				AnchorPoint={new Vector2(0.5, 0.5)}
+				BackgroundTransparency={1}
+				Position={new UDim2(0.5, 0, 0.5, 0)}
+				Size={new UDim2(1, 0, 1, 0)}
+				{...Native}
+			>
+				{children}
+			</frame>
+		);
+	},
+);
