+++ src/client/ui/components/primitive/index.ts
@@ -1,0 +1,9 @@
+export * from "./button";
+export * from "./canvas-group";
+export * from "./frame";
+export * from "./group";
+export * from "./image";
+export * from "./layer";
+export * from "./padding-component";
+export * from "./scrolling-frame";
+export * from "./text";
