+++ src/client/ui/components/ultra-wide-container.tsx
@@ -1,7 +1,7 @@
 import { useViewport } from "@rbxts/pretty-react-hooks";
 import React from "@rbxts/react";
 
-import Group from "./primitive/group";
+import { Group } from "./primitive";
 
 const MAX_ASPECT_RATIO = 19 / 9;
 const BASE_RESOLUTION = new Vector2(1920, 1020);
@@ -18,7 +18,7 @@ type UltraWideContainerProps = React.PropsWithChildren;
  * @returns The rendered container component.
  * @see https://github.com/Quenty/NevermoreEngine/tree/a9256cab3584bea4bd32c327d00b9a52f2a3ec95/src/ultrawidecontainerutils
  */
-export default function UltraWideContainer({
+export function UltraWideContainer({
 	children,
 }: Readonly<UltraWideContainerProps>): React.ReactNode {
 	const viewport = useViewport();
